refactor(mobile): break down MasterPasswordDialog into modular components
